Create Class_level_attribute.py
Class-level attributes
Class attributes store data that is shared among all the class instances. They are assigned values in the class body, and are referred to using the ClassName. syntax rather than self. syntax when used in methods.

In this exercise, you will be a game developer working on a game that will have several players moving on a grid and interacting with each other. As the first step, you want to define a Player class that will just move along a straight line. Player will have a position attribute and a move() method. The grid is limited, so the position of Player will have a maximal value.


Instruction:

Add a move() method with a steps parameter such that:

if position plus steps is less than MAX_POSITION, then add steps to position and assign the result back to position;
otherwise, set position to MAX_POSITION.